- Coalesce parquet into a single file - Saperate Token frequency filter into it's own function
